+++ src/main/java/com/rollingpaper/ggeujeogggeujeog/common/error/BaseException.java
@@ -1,4 +1,4 @@
-package com.rollingpaper.ggeujeogggeujeog.common.exception;
+package com.rollingpaper.ggeujeogggeujeog.common.error;
 
 import org.springframework.http.HttpStatus;
