+++ src/main/java/com/rollingpaper/ggeujeogggeujeog/common/error/ExceptionController.java
@@ -1,4 +1,4 @@
-package com.rollingpaper.ggeujeogggeujeog.common.exception;
+package com.rollingpaper.ggeujeogggeujeog.common.error;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
