--- src/main/java/com/rollingpaper/ggeujeogggeujeog/event/domain/EventType.java
@@ -1,7 +1,0 @@
-package com.rollingpaper.ggeujeogggeujeog.event.domain;
-
-public enum EventType {
-	INSERT,
-	UPDATE,
-	DELETE
-}
